+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## v3.2.0
+### Added
+- Added AggregatePNL class
+
+### Changed
+- Retrieve aggregate pnl returns class
+
 ## v3.1.0 - 2018-01-12
 ### Added
 - Added various EOD price APIs